validate restricted firlds

+++ src/controllers/medicineController.js
@@ -1,6 +1,5 @@
 import Medicine from "../models/medicine.js";
-import Pharmacy from "../models/pharmacy.js";
-import { validateUser } from "../service/commonService.js";
+import { validateUser, validateEditFields } from "../service/commonService.js";
 
 export const createMedicine = async (req, res) => {
   const { userId, email, role } = req.user;
@@ -17,7 +16,10 @@ export const createMedicine = async (req, res) => {
   } = req.body;
 
   try {
-    await validateUser(userId, email, role);
+    const user = validateUser(userId, email, role);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
 
     const newMedicine = new Medicine({
       pharmacyId: userId,
@@ -37,7 +39,9 @@ export const createMedicine = async (req, res) => {
     return res.status(201).json(newMedicine);
   } catch (error) {
     if (error.code === 11000) {
-      return res.status(409).json({ message: "Identification code must be unique." });
+      return res
+        .status(409)
+        .json({ message: "Identification code must be unique." });
     }
     console.error(error);
     return res.status(500).json({ message: "Error creating medicine" });
@@ -50,10 +54,17 @@ export const editMedicine = async (req, res) => {
   const updatedData = req.body;
 
   try {
-    await validateUser(userId, email, role);
+    const user = validateUser(userId, email, role);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
 
-    if(updatedData.identificationCode){
-      return res.status(400).json({ message: "Identification code cannot be updated." });
+    const noRestrictedFirlds = validateEditFields(updatedData);
+
+    if (!noRestrictedFirlds) {
+      return res
+        .status(400)
+        .json({ message: "Cannot update restricted fields" });
     }
 
     const errors = await validateMedicineData(updatedData);
@@ -83,7 +94,10 @@ export const deleteMedicine = async (req, res) => {
   const { medicineId } = req.params;
 
   try {
-    await validateUser(userId, email, role);
+    const user = validateUser(userId, email, role);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
 
     const deletedMedicine = await Medicine.findByIdAndDelete(medicineId);
 
@@ -103,7 +117,10 @@ export const getMedicineById = async (req, res) => {
   const { medicineId } = req.params;
 
   try {
-    await validateUser(userId, email, role);
+    const user = validateUser(userId, email, role);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
 
     const medicine = await Medicine.findById(medicineId).populate(
       "pharmacyId",
@@ -124,8 +141,10 @@ export const getAllMedicines = async (req, res) => {
   const { pharmacyId } = req.params;
 
   try {
-    await validateUser(userId, email, role);
-
+    const user = validateUser(userId, email, role);
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
     const medicines = await Medicine.find({ pharmacyId }).populate(
       "pharmacyId",
       "name email"
